Move tests from frontend.py into test_frontend.py
