+++ frontend.py
@@ -6,8 +6,6 @@ import json
 import queue
 import socketserver
 import threading
-import unittest
-from unittest.mock import MagicMock
 
 import mailqueue
 
@@ -16,7 +14,7 @@ incoming_queue = queue.Queue(10)
 
 
 class ThreadedHTTPServer(socketserver.ThreadingMixIn, http.server.HTTPServer):
-    "This server spawns a new thread for each incoming request"
+    """This server spawns a new thread for each incoming request"""
     pass
 
 
@@ -50,9 +48,8 @@ class Handler(http.server.BaseHTTPRequestHandler):
         self.end_headers()
         self.wfile.write('{}\n'.format(json.dumps(body)).encode())
 
-###################################################################################################
 
-class SendHandler():
+class SendHandler:
     def __init__(self, incoming_queue):
         self._incoming_queue = incoming_queue
 
@@ -68,78 +65,6 @@ class SendHandler():
 
         self._incoming_queue.put(message)
 
-
-class TestSendHandler(unittest.TestCase):
-    def setUp(self):
-        self.q = queue.Queue(1)
-        self.handler = SendHandler(self.q)
-        self.valid_request_dict = {
-            'sender': 'someone',
-            'recipients': 'alice, bob',
-            'subject': 'important message',
-            'body': 'hello!'
-        }
-
-    def test_valid_request_should_place_message_to_incoming_queue(self):
-        self.q.put = MagicMock()
-        self.handler.run(self.valid_request_dict)
-        self.q.put.assert_called_once()
-
-    def test_missing_request_keys_should_not_affect_queue(self):
-        for k in self.valid_request_dict:
-            with self.subTest(field=k):
-                q = queue.Queue(1)
-                q.put = MagicMock()
-                handler = SendHandler(q)
-
-                r = dict(self.valid_request_dict)
-                del r[k]
-
-                with self.assertRaises(ValueError) as cm:
-                    handler.run(r)
-
-                self.assertTrue(k in cm.exception.args[0],
-                                msg='expected exception description "{}" to contain "{}"'.format(
-                                    cm.exception.args[0], k
-                                ))
-                q.put.assert_not_called()
-
-    def test_message_body_should_be_taken_from_request(self):
-        expected_body = 'this will be a body of the message'
-        r = dict(self.valid_request_dict)
-        r['body'] = expected_body
-
-        self.handler.run(r)
-
-        message = self._queue_pop()
-        self.assertEqual(expected_body + '\n', message.get_content())
-
-    def test_message_header_should_contain_values_from_request(self):
-        input_fields = {
-            'sender': 'From',
-            'recipients': 'To',
-            'subject': 'Subject'
-        }
-
-        test_value = 'value_used_by_test'
-
-        for input_field, message_field in input_fields.items():
-            with self.subTest(input_field=input_field, message_field=message_field):
-                r = dict(self.valid_request_dict)
-                r[input_field] = test_value
-
-                self.handler.run(r)
-
-                message = self._queue_pop()
-                self.assertEqual(test_value, message[message_field])
-
-    def _queue_pop(self):
-        self.assertFalse(self.q.empty(),
-                         msg='the queue is empty, expected it to contain at least one message')
-        return self.q.get()
-
-
-###################################################################################################
 
 class MailQueueAppender(threading.Thread):
     def __init__(self, input_queue, outq_factory):
@@ -165,28 +90,6 @@ class MailQueueAppender(threading.Thread):
         message = self._input_queue.get()
         self._mailqueue.put(message)
 
-
-class TestMailQueueAppender(unittest.TestCase):
-    def test_happy_path(self):
-        # Using the "Humble Object" pattern to test the logic which is executed in a separate
-        # thread in production code: http://xunitpatterns.com/Humble%20Object.html
-        #
-        # Therefore we have to be a little more invasive than usually (e.g. we are accessing a
-        # private method).
-        input_queue = queue.Queue(1)
-        input_queue.put('something')
-
-        def outq_factory():
-            q = mailqueue.MailQueue()
-            q.put = MagicMock()
-            return q
-
-        appender = MailQueueAppender(input_queue, outq_factory)
-        appender._relay_one_message()
-
-        appender._mailqueue.put.assert_called_once()
-
-###################################################################################################
 
 def main():
     mailq_appender = MailQueueAppender(incoming_queue, mailqueue.MailQueue)
